+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 NEXT_VERSION_BUMP: MINOR
 ## XX.XX.XX - 20XX-XX-XX
 
+### Payments
+* [ADDED] Added Klarna confirmation options for passing SDK-generated interoperability and partner confirmation tokens to PaymentIntent confirmation (private preview).
+
 ### PaymentSheet
 * [CHANGED] Inline address autocomplete is now enabled by default in PaymentSheet and FlowController.